+++ src/gcovr/formats/gcov/parser/text.py
@@ -370,12 +370,59 @@ def _reconstruct_source_code(tokens: Iterable[_Line]) -> list[str]:
 
 
 class _ParserState(NamedTuple):
+    """State information while parsing gcov lines.
+
+    >>> state = _ParserState()
+    >>> state.previous_state is None
+    True
+    >>> state = state._replace(linecov_list=["x"])
+    >>> state.linecov_list == ["x"]
+    True
+    >>> new_state = state.save_state()
+    >>> new_state == state
+    False
+    >>> new_state.previous_state == state
+    True
+    >>> new_state.linecov_list == ["x"]
+    True
+    >>> state.linecov_list == ["x"]
+    True
+    >>> state = new_state._replace(linecov_list=[])
+    >>> state.linecov_list == []
+    True
+    >>> restored_state = state.restore_state()
+    >>> restored_state.previous_state is None
+    True
+    >>> state.linecov_list == []
+    True
+    >>> restored_state.restore_state()
+    Traceback (most recent call last):
+    ...
+    RuntimeError: Sanity check failed, previous_state of _ParserState is None.
+    """
+
     deferred_functions: list[_FunctionLine] = []
     function_name: Optional[str] = None
     function_specialization: bool = False
     linecov_list: list[LineCoverage] = []
     block_id: Optional[int] = None
     is_recovering: bool = False
+    previous_state: Optional["_ParserState"] = None
+
+    def save_state(self) -> "_ParserState":
+        """Save the current parser state and return a new one with the current line coverage list."""
+        return _ParserState(
+            previous_state=self,
+            linecov_list=self.linecov_list,
+        )
+
+    def restore_state(self) -> "_ParserState":
+        """Restore the previous parser state with the current line coverage list."""
+        if self.previous_state is None:
+            raise RuntimeError(
+                f"Sanity check failed, previous_state of {type(self).__name__} is None."
+            )
+        return self.previous_state._replace(linecov_list=self.linecov_list)
 
 
 def _gather_coverage_from_line(
@@ -427,7 +474,7 @@ def _gather_coverage_from_line(
                     if linecov.lineno >= lineno
                 ]
                 LOGGER.debug(
-                    f"Removing line coverage for line {to_remove[0].lineno} {'' if len(to_remove) == 1 else (' to line ' + str(to_remove[-1].lineno))} from previous function."
+                    f"Removing line coverage for line {to_remove[0].lineno}{'' if len(to_remove) == 1 else (' to line ' + str(to_remove[-1].lineno))} from previous function."
                 )
                 for linecov in to_remove:
                     filecov.remove_line_coverage(linecov)
@@ -471,18 +518,18 @@ def _gather_coverage_from_line(
             function_name=line.name,
         )
 
-    # currently, the parser just ignores specialization sections
     elif isinstance(line, _FunctionSpecializationNameLine):
-        return state
-
-    elif isinstance(line, _FunctionSpecializationSeparatorLine):
-        # Defer handling of the function tag until the next source line.
-        # This is important to get correct function name information.
-        return state._replace(
-            deferred_functions=[],
-            function_name=None,
+        # Now we know that we are in a specialization and not at the end of it.
+        return state.save_state()._replace(
             function_specialization=True,
         )
+
+    elif isinstance(line, _FunctionSpecializationSeparatorLine):
+        # If there was a specialization active we need to restore the previous state, else we do nothing.
+        if state.function_specialization:
+            state = state.restore_state()
+
+        return state
 
     elif isinstance(line, _BranchLine):
         branchno, hits, annotation = line