+++ tests/use-existing/test_use_existing.py
@@ -1,7 +1,6 @@
-import typing
+import pytest
 
-if typing.TYPE_CHECKING:
-    from tests.conftest import GcovrTestExec
+from tests.conftest import IS_GCC, GcovrTestExec
 
 
 def test_all(gcovr_test_exec: "GcovrTestExec", check) -> None:  # type: ignore[no-untyped-def]
@@ -27,7 +26,7 @@ def test_all(gcovr_test_exec: "GcovrTestExec", check) -> None:  # type: ignore[n
     )
     check.is_true(
         list(gcovr_test_exec.output_dir.glob("*.gcda")),
-        "*.gcda files should be touched",
+        "*.gcda files should not be touched",
     )
     check.is_true(
         list(gcovr_test_exec.output_dir.glob("*.gcov")), "*.gcov files should be kept"
@@ -65,5 +64,32 @@ def test_exclude_existing(gcovr_test_exec: "GcovrTestExec", check) -> None:  # t
         "--gcov-delete",
         "--json-pretty",
         "--json=coverage.json",
+    )
+    gcovr_test_exec.compare_json()
+
+
+@pytest.mark.skipif(
+    not IS_GCC,
+    reason="We use an existing file and this is independent from compiler",
+)
+def test_issue_1168(gcovr_test_exec: "GcovrTestExec", check) -> None:  # type: ignore[no-untyped-def]
+    """A test that verifies exclusion when using existing *.gcov coverage files."""
+    gcov_file = (
+        gcovr_test_exec.output_dir
+        / "ParamWithIntArray_test.cpp.gcda.ParamWithValue.h.gcov"
+    )
+    gcov_file_content = gcov_file.read_text().replace(
+        "/home/abrown/Catena/sdks/cpp/common/include",  # cspell:disable-line
+        str(gcovr_test_exec.output_dir),
+    )
+    gcov_file.write_text(gcov_file_content)
+    gcovr_test_exec.gcovr(
+        "--verbose",
+        "--gcov-use-existing-files",
+        "--filter=.*ParamWithValue.*",
+        "--trace-include=.*ParamWithValue.*",
+        "--json-pretty",
+        "--json=coverage.json",
+        ".",
     )
     gcovr_test_exec.compare_json()